Updated footer to include two sections: footer and legal
Legal entails more of the mandatory notices or copyright stuff, such as
the privacy policy. The main footer has social media links or other
content related stuff. This allows for adding content in the middle of
these sections, as well as arranging these sections easier depending on
the desired site design.